[arccore+arcane] Move class 'HPReal' from Arcane to Arccore.
